remove initialPlan folder and move files in pageForAH
